% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dend.R
\name{Dend}
\alias{Dend}
\title{Dendrograma de asociaciones y características por grupo}
\usage{
Dend(BD, num.gpos, Lista, nombres, metod.ligamiento, metod.dist, cex.1, cex.2)
}
\arguments{
\item{BD=Base}{de datos del análisis en csv; num.gpos=núm. de grupos deseados para la clasificación; nombres=1: quitar nombres de descriptores del de las especies, 0:no hacerlo}
}
\description{
Esta función hace dendrograma de asociaciones de la vegetación y saca algunas características por grupo.importar datos con encoding utf-8, stringasFactor=F. Se basa en los archivos creados por veg_an
}
\examples{
Dend("db.csv",6,1)
Dend()
}
\keyword{clasificación}
\keyword{de}
\keyword{dendrograma,}
\keyword{vegetación}

