% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descripcion.R
\name{descripcion}
\alias{descripcion}
\title{Análisis de vegetación por estratos}
\usage{
descripcion(BD)
}
\arguments{
\item{BD=Base}{de datos del análisis en csv}
}
\description{
Esta función saca descripciones de los sitios que se encuentran en Sitios unique.csv
}
\examples{
descripcion("db.csv",5,"Arboreo")
descripcion()
}
\keyword{descripción}
\keyword{sitios}

