% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitios.R
\name{sitios}
\alias{sitios}
\title{Análisis de vegetación por estratos}
\usage{
sitios(Sites.grp, Lista)
}
\arguments{
\item{Sites.grp=Base}{de datos del análisis en csv}

\item{radio=radio}{del círculo de muestreo, p.ej. 0.5,3,5}

\item{estrato=Arboreo, }{Arbustivo, Bajo (debe de ir entre comillas)}
}
\description{
Esta función calcula orientaciones y intervalo de altitudes para los sitios de estudio
}
\examples{
sitios("Sitios.csv",5,"Arboreo")
sitios()
}
\keyword{análisis,}
\keyword{estratos}
\keyword{vegetación,}

