% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP.R
\name{SP}
\alias{SP}
\title{Curva de acumulación de especies}
\usage{
SP(numsp.est, num.sit.est, break.x, break.y, int.lowess)
}
\arguments{
\item{Lee}{los archivos creados por vegan}
}
\description{
Esta función calcula curva de acumulación de especies.importar datos con encoding utf-8, stringasFactor=F. Se basa en los archivos creados por veg_an
}
\examples{
SP()
SP()
}
\keyword{clasificación}
\keyword{de}
\keyword{dendrograma,}
\keyword{vegetación}

