% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_an.R
\name{veg_an}
\alias{veg_an}
\title{Análisis de vegetación por estratos}
\usage{
veg_an(BD, radio, parcela, estrato)
}
\arguments{
\item{BD=Base}{de datos del análisis en csv}

\item{radio=radio}{del círculo de muestreo, p.ej. 0.5,3,5}

\item{estrato=Arboreo, }{Arbustivo, Bajo (debe de ir entre comillas)}
}
\description{
Esta función calcula distintas características de la vegetación analizada calcula las características estructurales por sitio, por especie y por especie y sitio.importar datos con encoding utf-8, stringasFactor=F
}
\examples{
veg_an("db.csv",5,"Arboreo")
veg_an()
}
\keyword{análisis,}
\keyword{estratos}
\keyword{vegetación,}

